Move snapshot.js -> to .ts w/ passing specs.
